requirements.txt
pyyaml
numpy
aiohttp